Make NetworkInformation target ns1.7
While doing so I preserved the UWP build since it was significant.
To do that I needed to make both Win32.Primitives and InteropServices
expose their ns1.7 surface area to UAP10.1 via copying the impl.
