+++ src/System.Net.NetworkInformation/ref/project.json
@@ -1,14 +1,10 @@
 {
   "dependencies": {
-    "System.Net.Primitives": "4.0.10",
-    "System.Runtime": "4.0.0",
-    "System.Threading.Tasks": "4.0.0"
+    "System.Net.Primitives": "4.4.0-beta-24619-02",
+    "System.Runtime": "4.4.0-beta-24619-02",
+    "System.Threading.Tasks": "4.4.0-beta-24619-02"
   },
   "frameworks": {
-    "netstandard1.3": {
-      "imports": [
-        "dotnet5.4"
-      ]
-    }
+    "netstandard1.7": {}
   }
 }
