--- src/System.Net.NetworkInformation/src/netcore50/project.json
@@ -1,21 +1,0 @@
-{
-  "frameworks": {
-    "netcore50": {
-      "dependencies": {
-        "Microsoft.NETCore.Platforms": "1.0.1",
-        "Microsoft.TargetingPack.Private.WinRT": "1.0.1",
-        "Microsoft.Win32.Primitives": "4.0.0",
-        "System.Diagnostics.Tracing": "4.0.20",
-        "System.Linq": "4.0.0",
-        "System.Net.Sockets": "4.1.0",
-        "System.Net.Primitives": "4.0.10",
-        "System.Runtime.InteropServices.WindowsRuntime": "4.0.0",
-        "System.Threading.Tasks": "4.0.10"
-      }
-    }
-  },
-  "runtimes": {
-    "win8-x86": {},
-    "win8-x64": {}
-  }
-}
