+++ src/System.Net.NetworkInformation/src/project.json
@@ -1,29 +1,45 @@
 {
   "frameworks": {
-    "netstandard1.3": {
+    "netstandard1.7": {
       "dependencies": {
-        "Microsoft.NETCore.Platforms": "1.0.1",
-        "Microsoft.Win32.Primitives": "4.0.1",
-        "System.Diagnostics.Contracts": "4.0.0",
-        "System.Diagnostics.Tracing": "4.0.0",
-        "System.IO.FileSystem": "4.0.0",
-        "System.Linq": "4.0.0",
-        "System.Net.Sockets": "4.1.0",
-        "System.Runtime.InteropServices": "4.1.0",
-        "System.Security.Principal.Windows": "4.0.0",
-        "System.Threading": "4.0.11",
-        "System.Threading.Timer": "4.0.1",
-        "System.Threading.Thread": "4.0.0",
-        "System.Threading.ThreadPool": "4.0.10",
-        "System.Threading.Overlapped": "4.0.0"
-      },
-      "imports": [
-        "dotnet5.4"
-      ]
+        "Microsoft.Win32.Primitives": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Contracts": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Debug": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Tracing": "4.4.0-beta-24619-02",
+        "System.IO.FileSystem": "4.4.0-beta-24619-02",
+        "System.Linq": "4.4.0-beta-24619-02",
+        "System.Net.Sockets": "4.4.0-beta-24619-02",
+        "System.Resources.ResourceManager": "4.4.0-beta-24619-02",
+        "System.Runtime.Extensions": "4.4.0-beta-24619-02",
+        "System.Runtime.InteropServices": "4.4.0-beta-24619-02",
+        "System.Security.Principal.Windows": "4.4.0-beta-24619-02",
+        "System.Threading": "4.4.0-beta-24619-02",
+        "System.Threading.Timer": "4.4.0-beta-24619-02",
+        "System.Threading.Thread": "4.4.0-beta-24619-02",
+        "System.Threading.ThreadPool": "4.4.0-beta-24619-02",
+        "System.Threading.Overlapped": "4.4.0-beta-24619-02"
+      }
     },
-    "net46": {
+    "net463": {
       "dependencies": {
         "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
+      }
+    },
+    "uap10.1": {
+      "dependencies": {
+        "Microsoft.TargetingPack.Private.WinRT": "1.0.1",
+        "Microsoft.Win32.Primitives": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Debug": "4.4.0-beta-24619-02",
+        "System.Diagnostics.Tracing": "4.4.0-beta-24619-02",
+        "System.Linq": "4.4.0-beta-24619-02",
+        "System.Net.Sockets": "4.4.0-beta-24619-02",
+        "System.Net.Primitives": "4.4.0-beta-24619-02",
+        "System.Resources.ResourceManager": "4.4.0-beta-24619-02",
+        "System.Runtime.Extensions": "4.4.0-beta-24619-02",
+        "System.Runtime.InteropServices": "4.4.0-beta-24619-02",
+        "System.Runtime.InteropServices.WindowsRuntime": "4.4.0-beta-24619-02",
+        "System.Threading": "4.4.0-beta-24619-02",
+        "System.Threading.Tasks": "4.4.0-beta-24619-02"
       }
     }
   }
