+++ src/System.Net.NetworkInformation/tests/FunctionalTests/System.Net.NetworkInformation.Functional.Tests.csproj
@@ -11,7 +11,7 @@
     <OutputType>Library</OutputType>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
     <ResourcesSourceOutputDirectory>../../src/Resources/</ResourcesSourceOutputDirectory>
-    <NugetTargetMoniker>.NETStandard,Version=v1.3</NugetTargetMoniker>
+    <NugetTargetMoniker Condition="'$(TargetGroup)' == ''">.NETStandard,Version=v1.7</NugetTargetMoniker>
   </PropertyGroup>
   <!-- Help VS understand available configurations -->
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'" />
@@ -52,7 +52,7 @@
     <Compile Include="IPInterfacePropertiesTest_Windows.cs" />
     <Compile Include="MiscParsingTests.cs" />
     <Compile Include="MockMapTcpState.cs" />
-    <Compile Include="NetworkChangeTest.cs" />
+    <Compile Include="NetworkAddressChangedTests.cs" />
     <Compile Include="NetworkInterfaceBasicTest.cs" />
     <Compile Include="PhysicalAddressTest.cs" />
     <Compile Include="ConnectionsParsingTests.cs" />
@@ -79,6 +79,10 @@
     <Compile Include="$(CommonTestPath)\System\PlatformDetection.cs">
       <Link>Common\tests\System\PlatformDetection.cs</Link>
     </Compile>
+  </ItemGroup>
+  <ItemGroup Condition="'$(TargetGroup)' == ''">
+    <Compile Include="NetworkAvailabilityChangedTests.cs" />
+    <Compile Include="NetworkInterfaceIPv4Statistics.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\pkg\System.Net.NetworkInformation.pkgproj">
