+++ src/System.Net.NetworkInformation/tests/System.Net.NetworkInformation.Tests.builds
@@ -2,25 +2,33 @@
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
   <ItemGroup>
-    <Project Include="FunctionalTests\System.Net.NetworkInformation.Functional.Tests.csproj">
-      <OSGroup>Linux</OSGroup>
-    </Project>
-    <Project Include="FunctionalTests\System.Net.NetworkInformation.Functional.Tests.csproj">
-      <OSGroup>OSX</OSGroup>
-    </Project>
+    <Project Include="FunctionalTests\System.Net.NetworkInformation.Functional.Tests.csproj" />
     <Project Include="FunctionalTests\System.Net.NetworkInformation.Functional.Tests.csproj">
       <OSGroup>Windows_NT</OSGroup>
-      <TestTFMs>netcore50;netcoreapp1.0;net46</TestTFMs>
+      <TestTFMs>net463</TestTFMs>
+    </Project>
+    <Project Include="FunctionalTests\System.Net.NetworkInformation.Functional.Tests.csproj">
+      <TargetGroup>netstandard1.3</TargetGroup>
+      <TestTFMs>netcoreapp1.0</TestTFMs>
+    </Project>
+    <Project Include="FunctionalTests\System.Net.NetworkInformation.Functional.Tests.csproj">
+      <TargetGroup>netstandard1.3</TargetGroup>
+      <OSGroup>Windows_NT</OSGroup>
+      <TestTFMs>netcore50;net46</TestTFMs>
+    </Project>
+    <Project Include="UnitTests\System.Net.NetworkInformation.WinRT.Unit.Tests.csproj" />
+    <Project Include="UnitTests\System.Net.NetworkInformation.WinRT.Unit.Tests.csproj">
+      <OSGroup>Windows_NT</OSGroup>
+      <TestTFMs>net463</TestTFMs>
     </Project>
     <Project Include="UnitTests\System.Net.NetworkInformation.WinRT.Unit.Tests.csproj">
-      <OSGroup>Linux</OSGroup>
-    </Project>
-    <Project Include="UnitTests\System.Net.NetworkInformation.WinRT.Unit.Tests.csproj">
-      <OSGroup>OSX</OSGroup>
+      <TargetGroup>netstandard1.3</TargetGroup>
+      <TestTFMs>netcoreapp1.0</TestTFMs>
     </Project>
     <Project Include="UnitTests\System.Net.NetworkInformation.WinRT.Unit.Tests.csproj">
       <OSGroup>Windows_NT</OSGroup>
-      <TestTFMs>netcore50;netcoreapp1.0;net46</TestTFMs>
+      <TargetGroup>netstandard1.3</TargetGroup>
+      <TestTFMs>netcore50;net46</TestTFMs>
     </Project>
   </ItemGroup>
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.traversal.targets))\dir.traversal.targets" />
