+++ src/System.Net.NetworkInformation/tests/UnitTests/System.Net.NetworkInformation.WinRT.Unit.Tests.csproj
@@ -10,7 +10,7 @@
     <ProjectGuid>{2194D6A2-1A35-46B5-8233-AEEBCBD31EF9}</ProjectGuid>
     <OutputType>Library</OutputType>
     <StringResourcesPath>../../src/Resources/Strings.resx</StringResourcesPath>
-    <NugetTargetMoniker>.NETStandard,Version=v1.3</NugetTargetMoniker>
+    <NugetTargetMoniker Condition="'$(TargetGroup)' == ''">.NETStandard,Version=v1.7</NugetTargetMoniker>
   </PropertyGroup>
   <!-- Help VS understand available configurations -->
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'" />
