+++ src/System.Net.NetworkInformation/tests/UnitTests/project.json
@@ -1,6 +1,5 @@
 {
   "dependencies": {
-    "Microsoft.NETCore.Platforms": "1.2.0-beta-24619-02",
     "System.Diagnostics.Tracing": "4.4.0-beta-24619-02",
     "System.Linq.Expressions": "4.4.0-beta-24619-02",
     "System.Net.Primitives": "4.4.0-beta-24619-02",
@@ -15,7 +14,8 @@
     "Microsoft.DotNet.BuildTools.TestSuite": "1.0.0-prerelease-00920-01"
   },
   "frameworks": {
-    "netstandard1.3": {}
+    "netstandard1.3": {},
+    "netstandard1.7": {}
   },
   "supports": {
     "coreFx.Test.netcore50": {},
