+++ src/System.Runtime.InteropServices/src/MissingTypes.uap101aot.cs
@@ -1,0 +1,20 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+// Temporary type definitions to quiet GenFacades
+namespace System.IO
+{
+    public class UnmanagedMemoryAccessor { }
+    public class UnmanagedMemoryStream { }
+}
+
+namespace System.Runtime.CompilerServices
+{
+    public interface IUnknownConstantAttribute {}
+}
+
+namespace System.Runtime.InteropServices
+{
+    public class RuntimeEnvironment {}
+}
