+++ src/System.Runtime.InteropServices/src/project.json
@@ -8,6 +8,11 @@
         "dotnet5.8"
       ]
     },
+    "uap10.1": {
+      "dependencies": {
+        "Microsoft.TargetingPack.Private.NETNative": "1.1.0-beta-24607-01"
+      }
+    },
     "net463": {
       "dependencies": {
         "Microsoft.TargetingPack.NETFramework.v4.6": "1.0.1"
